ref-tests: make the osaka ref tests (#2493)
* ref-tests: make the osaka ref tests


* ras


* the most boring stuff ever - no lispifying is worse


* overflow short


* ras typing update


* fix: arg of oob call are 0 if gt 1024 for osaka modexp


* ras typo


* ras


* fix: cursor comments

---------

+++ tracer/arithmetization/src/main/java/net/consensys/linea/zktracer/module/rlptxn/cancun/phaseSection/DataPhaseSection.java
@@ -71,7 +71,7 @@ public class DataPhaseSection extends PhaseSection {
     // trace the limbs
     int zeros = tx.numberOfZeroBytesInPayload();
     int nonZeros = tx.numberOfNonzeroBytesInPayload();
-    final short ctMax = (short) (limbs.size() - 1);
+    final int ctMax = limbs.size() - 1;
     for (int ct = 0; ct <= ctMax; ct++) {
       final InstructionDataPricing currentLimb = limbs.get(ct);
       traceTransactionConstantValues(trace, tracedValues);
